New Feature: Clarify why queries match
Previously, matches of query terms against, e.g., goods and services or the verbal description of a mark were returned, but the matching field was not identified. To alleviate any confusion and to explain how to investigate the match further (i.e., by clicking through the linked mark), footnotes were added.

JSON.stringify was used to flatten the JSON output of the TMOG API.